python/requirements.txt
# Core
websockets>=12.0
pyyaml>=6.0
python-osc>=1.8.0

# Audio
sounddevice>=0.4.6
soundcard>=0.4.2
numpy>=1.24.0
webrtcvad>=2.0.10

# STT
faster-whisper>=1.0.0
ctranslate2>=4.0.0

# Translation
transformers>=4.35.0
sentencepiece>=0.1.99
torch>=2.0.0

# TTS
piper-tts>=1.2.0
onnxruntime>=1.16.0
edge-tts>=6.1.0
pyttsx3>=2.90  # Windows SAPI wrapper

# LLM
llama-cpp-python>=0.2.50
openai>=1.10.0
anthropic>=0.18.0
groq>=0.4.0
google-generativeai>=0.4.0

# VR
openvr>=1.26.0
pillow>=10.0.0

# RVC Voice Conversion
faiss-cpu>=1.7.3
librosa>=0.9.2,<0.11.0
soundfile>=0.12.1
pydub>=0.25.1
static-ffmpeg>=3.0
miniaudio>=1.59

# VOICEVOX Engine setup
py7zr>=1.0.0

# Utilities
tqdm>=4.66.0
huggingface-hub>=0.20.0
requests>=2.31.0
aiohttp>=3.9.0
keyring>=24.0.0
